final fixes

+++ Marlin/src/module/settings.cpp
@@ -3348,7 +3348,7 @@ void MarlinSettings::postprocess() {
     // live at the very end of the eeprom
     const uint16_t MarlinSettings::meshes_end = persistentStore.capacity() - 129;
 
-    uint32_t MarlinSettings::meshes_start_index() {
+    uint16_t MarlinSettings::meshes_start_index() {
       // Pad the end of configuration data so it can float up
       // or down a little bit without disrupting the mesh data
       return (datasize() + EEPROM_OFFSET + 32) & 0xFFF8;
